put toNum and toStr in utils.ts
