+++ client/src/lib/utils.ts
@@ -4,3 +4,15 @@ import { twMerge } from 'tailwind-merge'
 export function cn(...inputs: ClassValue[]) {
   return twMerge(clsx(inputs))
 }
+
+export const toNum = (v: unknown): number => {
+  if (typeof v === 'number') return v
+  if (typeof v === 'string') {
+    const n = parseFloat(v.replace(/,/g, ''))
+    return isNaN(n) ? 0 : n
+  }
+  return 0
+}
+
+export const toStr = (v: unknown): string =>
+  typeof v === 'string' ? v : String(v ?? '')
